Huge pedanticizing and warnings elimination

+++ xar/lib/arcmod.c
@@ -68,7 +68,7 @@ struct arcmod xar_arcmods[] = {
 int32_t xar_arcmod_archive(xar_t x, xar_file_t f, const char *file, const char *buffer, size_t len) {
 	int i;
 	int32_t ret;
-	for(i = 0; i < (sizeof(xar_arcmods)/sizeof(struct arcmod)); i++) {
+	for(i = 0; i < (int)(sizeof(xar_arcmods)/sizeof(struct arcmod)); i++) {
 		if( xar_arcmods[i].archive ) {
 			ret = xar_arcmods[i].archive(x, f, file, buffer, len);
 			if( ret < 0 ) {
@@ -92,7 +92,7 @@ int32_t xar_arcmod_archive(xar_t x, xar_file_t f, const char *file, const char *
 int32_t xar_arcmod_extract(xar_t x, xar_file_t f, const char *file, char *buffer, size_t len) {
 	int i;
 	int32_t ret;
-	for(i = 0; i < (sizeof(xar_arcmods)/sizeof(struct arcmod)); i++) {
+	for(i = 0; i < (int)(sizeof(xar_arcmods)/sizeof(struct arcmod)); i++) {
 		if( xar_arcmods[i].extract ) {
 			ret = xar_arcmods[i].extract(x, f, file, buffer, len);
 			if( ret < 0 ) {
